Add actionhelper methods
