+++ src/redux/actions/cabinActions.js
@@ -1,9 +1,29 @@
-import Axios from 'axios';
+// import Axios from 'axios';
 import baseURL from './baseURL';
-import { queryBuilder } from './actionHelper';
+import queryBuilder from './actionHelper';
 
 const fetchCabin = (id) => {
   queryBuilder.get(`/api/v1/cabins/${id}.json`, 'FETCH_CABIN');
+};
+
+const fetchCabins = () => {
+  queryBuilder.get(`${baseURL}/api/v1/cabins.json`, 'FETCH_CABINS');
+};
+
+const fetchFavourites = () => {
+  queryBuilder.get('/api/v1/favourites.json', 'FAVOURITE_CABIN');
+};
+
+const createCabin = (data) => {
+  queryBuilder.post(data, '/api/v1/cabins.json', 'SUCCESS_MESSAGE');
+};
+
+const updateCabin = (data, id) => {
+  queryBuilder.updateCabin(data, `/api/v1/cabins/${id}.json`, 'UPDATE_CABIN');
+};
+
+const deleteCabin = (id) => {
+  queryBuilder.deletes(`/api/v1/cabins/${id}.json`, 'DELETE_CABIN');
 };
 
 //   Fetch Cabin
@@ -31,150 +51,150 @@ const fetchCabin = (id) => {
 // };
 
 //   Fetch Favourites
-const fetchFavourites = () => (dispatch) => {
-  const token = localStorage.getItem('jwt');
-  const authorizationAxios = Axios.create({
-    baseURL: `${baseURL}`,
-    headers: {
-      Authorization: `Bearer ${token}`,
-    },
-  });
+// const fetchFavourites = () => (dispatch) => {
+//   const token = localStorage.getItem('jwt');
+//   const authorizationAxios = Axios.create({
+//     baseURL: `${baseURL}`,
+//     headers: {
+//       Authorization: `Bearer ${token}`,
+//     },
+//   });
 
-  authorizationAxios
-    .get('/api/v1/favourites.json')
-    .then((res) => dispatch({
-      type: 'FAVOURITE_CABIN',
-      payload: res.data,
-    }))
-    .catch((err) => {
-      dispatch({
-        type: 'CREATE_ERROR',
-        payload: err,
-      });
-    });
-};
+//   authorizationAxios
+//     .get('/api/v1/favourites.json')
+//     .then((res) => dispatch({
+//       type: 'FAVOURITE_CABIN',
+//       payload: res.data,
+//     }))
+//     .catch((err) => {
+//       dispatch({
+//         type: 'CREATE_ERROR',
+//         payload: err,
+//       });
+//     });
+// };
 
 //   Fetch All Cabins
-const fetchCabins = () => (dispatch) => {
-  const token = localStorage.getItem('jwt');
-  const authorizationAxios = Axios.create({
-    baseURL: `${baseURL}`,
-    headers: {
-      Authorization: `Bearer ${token}`,
-    },
-  });
-  const url = `${baseURL}/api/v1/cabins.json`;
-  authorizationAxios.get(url)
-    .then((res) => dispatch({
-      type: 'FETCH_CABINS',
-      payload: res.data,
-    }))
-    .catch((err) => dispatch({
-      type: 'CREATE_ERROR',
-      payload: err,
-    }));
-};
+// const fetchCabins = () => (dispatch) => {
+//   const token = localStorage.getItem('jwt');
+//   const authorizationAxios = Axios.create({
+//     baseURL: `${baseURL}`,
+//     headers: {
+//       Authorization: `Bearer ${token}`,
+//     },
+//   });
+//   const url = `${baseURL}/api/v1/cabins.json`;
+//   authorizationAxios.get(url)
+//     .then((res) => dispatch({
+//       type: 'FETCH_CABINS',
+//       payload: res.data,
+//     }))
+//     .catch((err) => dispatch({
+//       type: 'CREATE_ERROR',
+//       payload: err,
+//     }));
+// };
 
 // Create Cabin
-const createCabin = (data) => (dispatch) => {
-  const token = localStorage.getItem('jwt');
-  const authAxios = Axios.create({
-    baseURL: `${baseURL}`,
-    headers: {
-      Authorization: `Bearer ${token}`,
-    },
-  });
-  const cabinData = {
-    cabin: data,
-  };
-  const payload = {
-    message: 'You have Successfully created a Cabin',
-    type: 'create_cabin',
-  };
-  authAxios
-    .post('/api/v1/cabins.json', cabinData)
-    .then((res) => {
-      dispatch({
-        type: 'SUCCESS_MESSAGE',
-        payload,
-      });
-      dispatch({
-        type: 'FETCH_CABIN',
-        payload: res.data,
-      });
-    })
-    .catch((err) => dispatch({
-      type: 'CREATE_ERROR',
-      payload: err,
-    }));
-};
+// const createCabin = (data) => (dispatch) => {
+//   const token = localStorage.getItem('jwt');
+//   const authAxios = Axios.create({
+//     baseURL: `${baseURL}`,
+//     headers: {
+//       Authorization: `Bearer ${token}`,
+//     },
+//   });
+//   const cabinData = {
+//     cabin: data,
+//   };
+//   const payload = {
+//     message: 'You have Successfully created a Cabin',
+//     type: 'create_cabin',
+//   };
+//   authAxios
+//     .post('/api/v1/cabins.json', cabinData)
+//     .then((res) => {
+//       dispatch({
+//         type: 'SUCCESS_MESSAGE',
+//         payload,
+//       });
+//       dispatch({
+//         type: 'FETCH_CABIN',
+//         payload: res.data,
+//       });
+//     })
+//     .catch((err) => dispatch({
+//       type: 'CREATE_ERROR',
+//       payload: err,
+//     }));
+// };
 
 // Update Cabin
-const updateCabin = (data, id) => (dispatch) => {
-  const token = localStorage.getItem('jwt');
-  const authAxios = Axios.create({
-    baseURL: `${baseURL}`,
-    headers: {
-      Authorization: `Bearer ${token}`,
-    },
-  });
-  const cabinData = {
-    cabin: data,
-  };
-  const payload = {
-    message: 'You have successfully updated a Cabin!',
-    type: 'update_cabin',
-  };
-  authAxios
-    .patch(`/api/v1/cabins/${id}.json`, cabinData)
-    .then((res) => {
-      dispatch({
-        type: 'SUCCESS_MESSAGE',
-        payload,
-      });
-      dispatch({
-        type: 'UPDATE_CABIN',
-        payload: res.data,
-      });
-    })
-    .catch((err) => dispatch({
-      type: 'CREATE_ERROR',
-      payload: err,
-    }));
-};
+// const updateCabin = (data, id) => (dispatch) => {
+//   const token = localStorage.getItem('jwt');
+//   const authAxios = Axios.create({
+//     baseURL: `${baseURL}`,
+//     headers: {
+//       Authorization: `Bearer ${token}`,
+//     },
+//   });
+//   const cabinData = {
+//     cabin: data,
+//   };
+//   const payload = {
+//     message: 'You have successfully updated a Cabin!',
+//     type: 'update_cabin',
+//   };
+//   authAxios
+//     .patch(`/api/v1/cabins/${id}.json`, cabinData)
+//     .then((res) => {
+//       dispatch({
+//         type: 'SUCCESS_MESSAGE',
+//         payload,
+//       });
+//       dispatch({
+//         type: 'UPDATE_CABIN',
+//         payload: res.data,
+//       });
+//     })
+//     .catch((err) => dispatch({
+//       type: 'CREATE_ERROR',
+//       payload: err,
+//     }));
+// };
 
 //   Delete Cabin
-const deleteCabin = (id) => (dispatch) => {
-  const token = localStorage.getItem('jwt');
+// const deleteCabin = (id) => (dispatch) => {
+//   const token = localStorage.getItem('jwt');
 
-  const authAxios = Axios.create({
-    baseURL: `${baseURL}`,
-    headers: {
-      Authorization: `Bearer ${token}`,
-    },
-  });
-  const payload = {
-    message: 'You have successfully deleted the cabin.',
-    type: 'delete_cabin',
-  };
-  authAxios
-    .delete(`/api/v1/cabins/${id}.json`)
-    .then((res) => {
-      dispatch({
-        type: 'SUCCESS_MESSAGE',
-        payload,
-      });
+//   const authAxios = Axios.create({
+//     baseURL: `${baseURL}`,
+//     headers: {
+//       Authorization: `Bearer ${token}`,
+//     },
+//   });
+//   const payload = {
+//     message: 'You have successfully deleted the cabin.',
+//     type: 'delete_cabin',
+//   };
+//   authAxios
+//     .delete(`/api/v1/cabins/${id}.json`)
+//     .then((res) => {
+//       dispatch({
+//         type: 'SUCCESS_MESSAGE',
+//         payload,
+//       });
 
-      dispatch({
-        type: 'DELETE_CABIN',
-        payload: res.data,
-      });
-    })
-    .catch((err) => dispatch({
-      type: 'CREATE_ERROR',
-      payload: err,
-    }));
-};
+//       dispatch({
+//         type: 'DELETE_CABIN',
+//         payload: res.data,
+//       });
+//     })
+//     .catch((err) => dispatch({
+//       type: 'CREATE_ERROR',
+//       payload: err,
+//     }));
+// };
 
 const unLoad = (states) => (dispatch) => {
   dispatch({
